@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import { useLocation, useHistory, Link } from "react-router-dom";
+import { useLocation, useNavigate, Link } from "react-router-dom";
 import { useTheme } from '@mui/material/styles';
 import { makeStyles, withStyles } from '@mui/styles';
 import { Typography, Container, Button, Paper, Box, Snackbar, Tooltip, TextField, CircularProgress, Card, CardContent, CardHeader, IconButton, MenuItem } from "@mui/material";
@@ -24,7 +24,7 @@ import { settings } from '../settings'
 
 export default function Evaluation() {
   const theme = useTheme();
-  const history = useHistory();
+  const navigate = useNavigate();
   // const auth = useAuth();
 
   const useStyles = makeStyles(() => ({
@@ -130,35 +130,35 @@ export default function Evaluation() {
     }
 
     // Get the list of evaluations from API
-    if (state.evaluationsList.length < 1) {
-      axios.get(settings.restUrl + '/evaluations', {
-        headers: {'Content-Type': 'application/json'},
-      })
-        .then((res: any) => {
-          let evaluationsList: any = []
-          res.data.map((evaluation: any, key: number) => {
-            evaluation['id'] = evaluation['_id']
-            evaluation['score_percent'] = evaluation['score_percent']
-            evaluation['score'] = evaluation['score'] + '/' + evaluation['score_max']
-            evaluation['subject'] = evaluation['subject']
-            evaluation['collection'] = evaluation['collection']
-            evaluation['created'] = evaluation['created_at']
-            evaluation['author'] = evaluation['author']
-            // evaluation['bonus_percent'] = evaluation['bonus']
-            evaluation['bonus_percent'] = 0
-            evaluationsList.push(evaluation)
-          })
-          updateState({ evaluationsList: evaluationsList })
-          updateState({ evaluationRunning: false })
-        })
-        .catch(function (error) {
-          updateState({
-            openError: 'inline',
-            evaluationRunning: false,
-            errorMessage: 'Error when retrieving the evaluations, please retry. If it persists, it means the server is probably having some issues! 😱'
-          })
-        })
-    }
+    // if (state.evaluationsList.length < 1) {
+    //   axios.get(settings.restUrl + '/evaluations', {
+    //     headers: {'Content-Type': 'application/json'},
+    //   })
+    //     .then((res: any) => {
+    //       let evaluationsList: any = []
+    //       res.data.map((evaluation: any, key: number) => {
+    //         evaluation['id'] = evaluation['_id']
+    //         evaluation['score_percent'] = evaluation['score_percent']
+    //         evaluation['score'] = evaluation['score'] + '/' + evaluation['score_max']
+    //         evaluation['subject'] = evaluation['subject']
+    //         evaluation['collection'] = evaluation['collection']
+    //         evaluation['created'] = evaluation['created_at']
+    //         evaluation['author'] = evaluation['author']
+    //         // evaluation['bonus_percent'] = evaluation['bonus']
+    //         evaluation['bonus_percent'] = 0
+    //         evaluationsList.push(evaluation)
+    //       })
+    //       updateState({ evaluationsList: evaluationsList })
+    //       updateState({ evaluationRunning: false })
+    //     })
+    //     .catch(function (error) {
+    //       updateState({
+    //         openError: 'inline',
+    //         evaluationRunning: false,
+    //         errorMessage: 'Error when retrieving the evaluations, please retry. If it persists, it means the server is probably having some issues! 😱'
+    //       })
+    //     })
+    // }
   }, [])
 
   const colors: any = {
@@ -185,71 +185,72 @@ export default function Evaluation() {
   }
 
   const doEvaluateUrl  = (evaluateUrl: string) => {
-    updateState({
-      evaluationRunning: true,
-      evaluationResults: null,
-      openError: 'none'
-    })
-    console.log('Starting evaluation of ' + evaluateUrl + ' with API ' + settings.docsUrl)
-    const postJson: any = {
-      "subject": evaluateUrl,
-      // "title": "FAIR metrics dataset evaluation",
-      "collection": state.collectionSelected
-    };
-    console.log(user)
-    // Getting User access_token from context set by NavBar
-    let headers: any = {'Content-Type': 'application/json'}
-    if (user && user['access_token']) {
-      headers['Authorization'] = 'Bearer ' + user['access_token']
-    }
-    axios.post(settings.restUrl + '/evaluations', JSON.stringify(postJson),
-      { headers: headers }
-    )
-      .then(res => {
-        axiosRetry(axios, {
-          retries: 30, // number of retries
-          retryDelay: (retryCount) => {
-            console.log(`Waiting for evaluation to finish: ${retryCount}`);
-            return retryCount * 3000; // time interval between retries
-          },
-          retryCondition: (error: any) => {
-            // if retry condition is not specified, by default idempotent requests are retried
-            return error.response.status === 404;
-          },
-        });
+    navigate("/evaluations?uri=" + evaluateUrl);
+    // updateState({
+    //   evaluationRunning: true,
+    //   evaluationResults: null,
+    //   openError: 'none'
+    // })
+    // console.log('Starting evaluation of ' + evaluateUrl + ' with API ' + settings.docsUrl)
+    // const postJson: any = {
+    //   "subject": evaluateUrl,
+    //   // "title": "FAIR metrics dataset evaluation",
+    //   "collection": state.collectionSelected
+    // };
+    // console.log(user)
+    // // Getting User access_token from context set by NavBar
+    // let headers: any = {'Content-Type': 'application/json'}
+    // if (user && user['access_token']) {
+    //   headers['Authorization'] = 'Bearer ' + user['access_token']
+    // }
+    // axios.post(settings.restUrl + '/evaluations', JSON.stringify(postJson),
+    //   { headers: headers }
+    // )
+    //   .then(res => {
+    //     axiosRetry(axios, {
+    //       retries: 30, // number of retries
+    //       retryDelay: (retryCount) => {
+    //         console.log(`Waiting for evaluation to finish: ${retryCount}`);
+    //         return retryCount * 3000; // time interval between retries
+    //       },
+    //       retryCondition: (error: any) => {
+    //         // if retry condition is not specified, by default idempotent requests are retried
+    //         return error.response.status === 404;
+    //       },
+    //     });
 
-        const evalId = res.data['_id']
-        // Retry every 3 seconds until the evaluation is available
-        axios.get(settings.restUrl + '/evaluations/' + evalId)
-          .then((res: any) => {
-            // Redirect to the page of the created evaluation
-            history.push("/evaluations/" + evalId);
-          })
-      })
-      .catch(function (error) {
-        updateState({
-          openError: 'inline',
-          evaluationRunning: false,
-          errorMessage: 'Error when running the evaluation, please retry.'
-        })
-        if (error.response) {
-          // Request made and server responded
-          // {"detail":[{"loc":["body","homepage"],"msg":"invalid or missing URL scheme","type":"value_error.url.scheme"}]}
-          if (error.response.data["detail"]) {
-            updateState({ errorMessage: 'Error: ' + JSON.stringify(error.response.data["detail"])})
-          } else {
-            updateState({ errorMessage: JSON.stringify(error.response.data) })
-          }
-        } else if (error.request) {
-          // The request was made but no response was received
-          console.log('request err');
-          console.log(error.request);
-        } else {
-          // Something happened in setting up the request that triggered an Error
-          console.log('Error', error.message);
-          updateState({ errorMessage: error.message })
-        }
-      })
+    //     const evalId = res.data['_id']
+    //     // Retry every 3 seconds until the evaluation is available
+    //     axios.get(settings.restUrl + '/evaluations/' + evalId)
+    //       .then((res: any) => {
+    //         // Redirect to the page of the created evaluation
+    //         navigate("/evaluations/" + evalId);
+    //       })
+    //   })
+    //   .catch(function (error) {
+    //     updateState({
+    //       openError: 'inline',
+    //       evaluationRunning: false,
+    //       errorMessage: 'Error when running the evaluation, please retry.'
+    //     })
+    //     if (error.response) {
+    //       // Request made and server responded
+    //       // {"detail":[{"loc":["body","homepage"],"msg":"invalid or missing URL scheme","type":"value_error.url.scheme"}]}
+    //       if (error.response.data["detail"]) {
+    //         updateState({ errorMessage: 'Error: ' + JSON.stringify(error.response.data["detail"])})
+    //       } else {
+    //         updateState({ errorMessage: JSON.stringify(error.response.data) })
+    //       }
+    //     } else if (error.request) {
+    //       // The request was made but no response was received
+    //       console.log('request err');
+    //       console.log(error.request);
+    //     } else {
+    //       // Something happened in setting up the request that triggered an Error
+    //       console.log('Error', error.message);
+    //       updateState({ errorMessage: error.message })
+    //     }
+    //   })
   }
 
   const handleTextFieldChange = (event: React.ChangeEvent<HTMLInputElement>) => {
@@ -279,94 +280,94 @@ export default function Evaluation() {
     // hljs.highlightAll();
   }
 
-  const columns: GridColumns = [
-    { field: '@id', headerName: 'ID', hide: true },
-    {
-      field: 'id', headerName: 'Access evaluation', flex: 0.4,
-      renderCell: (params: GridRenderCellParams) => (
-        <Link to={'/evaluations/' + params.value as string}>
-          <Button variant="contained"
-              className={classes.submitButton}
-              startIcon={<EvaluationIcon />}
-              color="primary" style={{textTransform: 'none'}}>
-            Results
-          </Button>
-        </Link>)
-    },
-    // { field: 'title', headerName: 'Title', flex: 1 },
-    {
-      field: 'subject', headerName: 'Resource URI', flex: 0.9,
-      renderCell: (params: GridRenderCellParams) => (
-        <>
-          {getUrlHtml(params.value as string)}
-        </>)
-    },
-    {
-      field: 'title', headerName: 'Resource title', flex: 0.9,
-      renderCell: (params: GridRenderCellParams) => (
-        <Tooltip title={params.value as string}>
-          <span>{params.value as string}</span>
-        </Tooltip>)
-    },
-    {
-      field: 'collection', headerName: 'Collection', flex: 0.5,
-      renderCell: (params: GridRenderCellParams) => (
-        <>
-          <Link to={'/collections/' + params.value as string} style={{textDecoration: 'none', color: theme.palette.primary.dark}}>
-            {params.value as string}
-          </Link>
-        </>)
-    },
-    {
-      field: 'score', headerName: 'Score', flex: 0.3,
-      renderCell: (params: GridRenderCellParams) => (
-        <>
-          {params.value as string}
-        </>)
-    },
-    {
-      field: 'score_percent', headerName: 'Completion', flex: 0.3,
-      renderCell: (params: GridRenderCellParams) => (
-        <>
-          {params.value as string}%
-          { params.value == 100 &&
-            <>
-              &nbsp;🔥
-            </>
-          }
-        </>)
-    },
-    // {
-    //   field: 'bonus_percent', headerName: 'Bonus score', flex: 0.3,
-    //   renderCell: (params: GridRenderCellParams) => (
-    //     <>
-    //       {params.value as string}%
-    //     </>)
-    // },
-    {
-      field: 'created', headerName: 'Date created', flex: 0.3,
-      renderCell: (params: GridRenderCellParams) => {
-        const dateCreated = params.value as string
-        return (
-        <>
-          {dateCreated.substring(0, dateCreated.indexOf('T'))}
-        </>)}
-    },
-    {
-      field: 'author', headerName: 'Author', flex: 0.6,
-      renderCell: (params: GridRenderCellParams) => (
-        <>
-          {getUrlHtml(params.value as string)}
-        </>)
-    }
-  ]
+  // const columns: GridColumns = [
+  //   { field: '@id', headerName: 'ID', hide: true },
+  //   {
+  //     field: 'id', headerName: 'Access evaluation', flex: 0.4,
+  //     renderCell: (params: GridRenderCellParams) => (
+  //       <Link to={'/evaluations/' + params.value as string}>
+  //         <Button variant="contained"
+  //             className={classes.submitButton}
+  //             startIcon={<EvaluationIcon />}
+  //             color="primary" style={{textTransform: 'none'}}>
+  //           Results
+  //         </Button>
+  //       </Link>)
+  //   },
+  //   // { field: 'title', headerName: 'Title', flex: 1 },
+  //   {
+  //     field: 'subject', headerName: 'Resource URI', flex: 0.9,
+  //     renderCell: (params: GridRenderCellParams) => (
+  //       <>
+  //         {getUrlHtml(params.value as string)}
+  //       </>)
+  //   },
+  //   {
+  //     field: 'title', headerName: 'Resource title', flex: 0.9,
+  //     renderCell: (params: GridRenderCellParams) => (
+  //       <Tooltip title={params.value as string}>
+  //         <span>{params.value as string}</span>
+  //       </Tooltip>)
+  //   },
+  //   {
+  //     field: 'collection', headerName: 'Collection', flex: 0.5,
+  //     renderCell: (params: GridRenderCellParams) => (
+  //       <>
+  //         <Link to={'/collections/' + params.value as string} style={{textDecoration: 'none', color: theme.palette.primary.dark}}>
+  //           {params.value as string}
+  //         </Link>
+  //       </>)
+  //   },
+  //   {
+  //     field: 'score', headerName: 'Score', flex: 0.3,
+  //     renderCell: (params: GridRenderCellParams) => (
+  //       <>
+  //         {params.value as string}
+  //       </>)
+  //   },
+  //   {
+  //     field: 'score_percent', headerName: 'Completion', flex: 0.3,
+  //     renderCell: (params: GridRenderCellParams) => (
+  //       <>
+  //         {params.value as string}%
+  //         { params.value == 100 &&
+  //           <>
+  //             &nbsp;🔥
+  //           </>
+  //         }
+  //       </>)
+  //   },
+  //   // {
+  //   //   field: 'bonus_percent', headerName: 'Bonus score', flex: 0.3,
+  //   //   renderCell: (params: GridRenderCellParams) => (
+  //   //     <>
+  //   //       {params.value as string}%
+  //   //     </>)
+  //   // },
+  //   {
+  //     field: 'created', headerName: 'Date created', flex: 0.3,
+  //     renderCell: (params: GridRenderCellParams) => {
+  //       const dateCreated = params.value as string
+  //       return (
+  //       <>
+  //         {dateCreated.substring(0, dateCreated.indexOf('T'))}
+  //       </>)}
+  //   },
+  //   {
+  //     field: 'author', headerName: 'Author', flex: 0.6,
+  //     renderCell: (params: GridRenderCellParams) => (
+  //       <>
+  //         {getUrlHtml(params.value as string)}
+  //       </>)
+  //   }
+  // ]
 
-  const [sortModel, setSortModel] = React.useState<GridSortModel>([
-    {
-      field: 'created',
-      sort: 'desc',
-    },
-  ]);
+  // const [sortModel, setSortModel] = React.useState<GridSortModel>([
+  //   {
+  //     field: 'created',
+  //     sort: 'desc',
+  //   },
+  // ]);
 
   return(
     <Container className='mainContainer'>
@@ -431,42 +432,8 @@ export default function Evaluation() {
       {/* Form to provide the URL to evaludate */}
       <form onSubmit={handleSubmit} style={{textAlign: 'left'}}>
 
-        <Typography variant='body1' style={{marginBottom: theme.spacing(3)}}>
-          1. Choose the collection of FAIR Maturity Indicators (aka. FAIR Metrics tests) to validate your resource against:
-        </Typography>
-
-        <Box display="flex" style={{marginBottom: theme.spacing(1)}}>
-          {/* Collection dropdown */}
-          <Box>
-            <TextField select
-                value={state.collectionSelected}
-                label={"Use the collection"}
-                id="collectionSelected"
-                onChange={handleCollectionChange}
-                SelectProps={{ MenuProps: { disableScrollLock: true } }}
-                style={{margin: theme.spacing(0, 0), backgroundColor: 'white'}}
-                variant="outlined">
-              { state.collectionsList.map((collec: any, key: number) => (
-                <MenuItem value={collec.id}>{collec.title} ({collec.id})</MenuItem>
-              ))}
-            </TextField>
-          </Box>
-
-          { state.collectionsList.map((collec: any, key: number) => {
-            if (collec.id === state.collectionSelected) {
-              return (
-                <Card style={{marginLeft: theme.spacing(1), padding: '15px'}}>
-                  <Typography variant='body1'>
-                    {collec.description}
-                  </Typography>
-                </Card>
-              )
-            }
-          })}
-        </Box>
-
         <Typography variant='body1' style={{marginBottom: theme.spacing(1)}}>
-          2. Provide the URL to the resource, or digital object, you want to evaluate <Tooltip
+          Provide the URL to the resource, or digital object, you want to evaluate <Tooltip
           title={<Typography>
             🔗 The best is to a URL, aka. URI, as it is the most standard and unambiguous protocol for identifying a resource online.
             <br/>But DOI and handles are also accepted by some collections
@@ -505,42 +472,6 @@ export default function Evaluation() {
         }
       </form>
 
-      <Box style={{margin: theme.spacing(4, 0)}}>
-        {state.evaluationRunning &&
-          <>
-            <Typography>
-              An evaluation can take up to 1 minute depending on the URL evaluated, and the collection used.
-            </Typography>
-            <Typography>
-              If you leave this page the evaluation will not stop, and you will be able to find it in the list below when it is done.
-            </Typography>
-            <CircularProgress style={{marginTop: '20px'}} />
-          </>
-        }
-        <Card elevation={4}
-            style={{background: "#e57373", padding: '15px', fontFamily: "Open Sans", fontSize: 12}}
-            sx={{ display: state.openError }}>
-          ⚠️&nbsp;&nbsp;{state.errorMessage}
-        </Card>
-      </Box>
-
-      {/* Display the Data table listing the Evaluations */}
-      {state.evaluationsList.length > 0 &&
-        <div style={{ height: 600, width: '100%' }}>
-          {/* {console.log(state.evaluationsList)} */}
-          <DataGrid
-            columns={columns}
-            rows={state.evaluationsList}
-            // {...state.evaluationsList}
-            components={{
-              Toolbar: GridToolbar,
-            }}
-            sortModel={sortModel}
-            onSortModelChange={(model) => setSortModel(model)}
-            style={{backgroundColor: '#fff'}}
-          />
-        </div>
-      }
     </Container>
   )
 }
